refactor(macos): split shell sidebar view owners
