+++ clients/apple/alan-macos/Views/Shell/ShellSidebarView.swift
@@ -879,1672 +879,4 @@ struct ShellSidebarView: View {
         }
     }
 }
-
-private struct ShellSidebarTabListOffsetPreferenceKey: PreferenceKey {
-    static var defaultValue: CGFloat = 0
-
-    static func reduce(value: inout CGFloat, nextValue: () -> CGFloat) {
-        value = nextValue()
-    }
-}
-
-private struct ShellSidebarTabDropTarget: Equatable {
-    let spaceID: String
-    let section: ShellTabOrganizationSection
-    let index: Int
-}
-
-private struct ShellSidebarTabDropDelegate: DropDelegate {
-    let target: ShellSidebarTabDropTarget
-    @Binding var activeDrag: ShellSidebarTabDragSource?
-    @Binding var preview: ShellSidebarTabDropTarget?
-    let host: ShellHostController
-
-    func dropEntered(info: DropInfo) {
-        preview = resolvedTarget(for: info)
-    }
-
-    func dropUpdated(info: DropInfo) -> DropProposal? {
-        preview = resolvedTarget(for: info)
-        return DropProposal(operation: .move)
-    }
-
-    func dropExited(info: DropInfo) {
-        if preview == target {
-            preview = nil
-        }
-    }
-
-    func performDrop(info: DropInfo) -> Bool {
-        let insertionTarget = resolvedTarget(for: info)
-        preview = nil
-
-        if let activeDrag {
-            self.activeDrag = nil
-            return performDrop(source: activeDrag, insertionTarget: insertionTarget)
-        }
-
-        self.activeDrag = nil
-        guard let provider = info.itemProviders(for: [.plainText]).first else {
-            return false
-        }
-        provider.loadObject(ofClass: NSString.self) { object, _ in
-            guard let payload = object as? String,
-                  let source = try? ShellSidebarTabDragSource.decodedPlainTextPayload(payload)
-            else {
-                return
-            }
-            DispatchQueue.main.async {
-                _ = performDrop(source: source, insertionTarget: insertionTarget)
-            }
-        }
-        return true
-    }
-
-    private func performDrop(
-        source: ShellSidebarTabDragSource,
-        insertionTarget: ShellSidebarTabDropTarget
-    ) -> Bool {
-        let activeDrag = source
-        let mutationIndex = mutationIndex(for: insertionTarget, activeDrag: activeDrag)
-
-        if source.sourceSpaceID == insertionTarget.spaceID,
-           source.sourceSection == insertionTarget.section,
-           source.sourceIndex == mutationIndex
-        {
-            return true
-        }
-
-        return host.reorderTab(
-            tabID: source.tabID,
-            targetSpaceID: insertionTarget.spaceID,
-            section: insertionTarget.section,
-            index: mutationIndex
-        )
-    }
-
-    private func mutationIndex(
-        for insertionTarget: ShellSidebarTabDropTarget,
-        activeDrag: ShellSidebarTabDragSource
-    ) -> Int {
-        ShellSidebarTabDropModel.mutationIndex(
-            for: ShellSidebarTabInsertionTarget(
-                spaceID: insertionTarget.spaceID,
-                section: insertionTarget.section,
-                index: insertionTarget.index
-            ),
-            source: activeDrag
-        )
-    }
-
-    private func resolvedTarget(for info: DropInfo) -> ShellSidebarTabDropTarget {
-        let rowMidpoint = ShellSidebarRowMetrics.dragMidpoint
-        let sectionCount = host.shellState
-            .space(spaceID: target.spaceID)?
-            .tabs(in: target.section)
-            .count ?? target.index
-        let adjustedIndex = info.location.y > rowMidpoint
-            ? target.index + 1
-            : target.index
-        return ShellSidebarTabDropTarget(
-            spaceID: target.spaceID,
-            section: target.section,
-            index: min(max(adjustedIndex, 0), sectionCount)
-        )
-    }
-}
-
-private struct ShellSidebarTabSectionDivider: View {
-    var body: some View {
-        Rectangle()
-            .fill(ShellPalette.line.opacity(0.30))
-            .frame(height: 0.5)
-            .padding(.horizontal, ShellSidebarMetrics.rowInset + 4)
-            .allowsHitTesting(false)
-    }
-}
-
-private struct ShellSidebarTabInsertionLine: View {
-    let isVisible: Bool
-
-    var body: some View {
-        RoundedRectangle(cornerRadius: ShellRadii.micro, style: .continuous)
-            .fill(ShellPalette.accent.opacity(isVisible ? 0.72 : 0))
-            .frame(height: isVisible ? 2 : 0)
-            .padding(.horizontal, ShellSidebarMetrics.rowInset + 2)
-            .padding(.vertical, isVisible ? 3 : 0)
-            .animation(.easeOut(duration: 0.10), value: isVisible)
-            .accessibilityHidden(true)
-    }
-}
-
-private struct ShellSidebarScrollBoundary: View {
-    let progress: CGFloat
-
-    var body: some View {
-        VStack(spacing: 0) {
-            Rectangle()
-                .fill(ShellPalette.line.opacity(0.36))
-                .frame(height: 0.5)
-
-            LinearGradient(
-                colors: [
-                    ShellPalette.sidebarInk.opacity(0.10),
-                    ShellPalette.sidebarInk.opacity(0.035),
-                    ShellPalette.sidebarInk.opacity(0),
-                ],
-                startPoint: .top,
-                endPoint: .bottom
-            )
-            .frame(height: 14)
-        }
-        .frame(maxWidth: .infinity, alignment: .top)
-        .opacity(progress)
-        .allowsHitTesting(false)
-    }
-}
-
-private struct ShellSidebarSpaceSliderScrollOffsetKey: PreferenceKey {
-    static var defaultValue: CGFloat = 0
-
-    static func reduce(value: inout CGFloat, nextValue: () -> CGFloat) {
-        value = nextValue()
-    }
-}
-
-/// Live descriptor for the in-progress Space being created. When present the
-/// slider appends one trailing, selected, display-only target reflecting the
-/// typed name/icon — no real Space exists until Create.
-struct ShellSpaceSliderDraft: Equatable {
-    let name: String
-    let iconSystemName: String?
-}
-
-private struct ShellSidebarSpaceSlider: View {
-    @Environment(\.accessibilityReduceMotion) private var reduceMotion
-    @ObservedObject var host: ShellHostController
-    let displaySpaceID: String?
-    let previewedSpaceID: String?
-    let activityFreshnessNow: Date
-    let onForwardVerticalWheel: (ShellSidebarSpaceSliderWheelEvent) -> Bool
-    var creationDraft: ShellSpaceSliderDraft? = nil
-    static let draftTargetID = "space-creation-draft"
-    @FocusState private var isKeyboardFocused: Bool
-    @State private var hoveredSpaceID: String?
-    @State private var scrubState: ShellSidebarSpaceSliderScrubState?
-    @State private var wheelIntentState = ShellSidebarSpaceSliderWheelIntentState()
-    @State private var wheelCommitToken = 0
-    @State private var trackScrollOffsetX: CGFloat = 0
-    @StateObject private var spaceCreationProfileOptions = ShellSpaceCreationProfileOptionStore()
-
-    var body: some View {
-        GeometryReader { proxy in
-            let spaces = visibleSpaces
-            let layout = sliderLayout(availableWidth: proxy.size.width)
-            let trackWidth = max(proxy.size.width - (ShellSidebarMetrics.edgeInset * 2), 0)
-
-            ScrollViewReader { scrollProxy in
-                ZStack(alignment: .leading) {
-                    RoundedRectangle(
-                        cornerRadius: ShellSidebarSpaceSliderLayout.trackHeight * 0.5,
-                        style: .continuous
-                    )
-                    .fill(ShellPalette.sidebarSpaceSliderTrack)
-                    .overlay {
-                        RoundedRectangle(
-                            cornerRadius: ShellSidebarSpaceSliderLayout.trackHeight * 0.5,
-                            style: .continuous
-                        )
-                        .stroke(ShellPalette.line.opacity(0.22), lineWidth: 0.6)
-                    }
-
-                    ScrollView(.horizontal, showsIndicators: false) {
-                        HStack(spacing: ShellSidebarSpaceSliderLayout.spacing) {
-                            ForEach(Array(spaces.enumerated()), id: \.element.spaceID) { index, space in
-                                if let item = layout.items.first(where: { $0.index == index }) {
-                                    spaceControl(for: space, item: item)
-                                        .id(space.spaceID)
-                                }
-                            }
-                            if let creationDraft,
-                               let draftItem = layout.items.first(where: { $0.index == spaces.count })
-                            {
-                                draftControl(for: creationDraft, item: draftItem)
-                                    .id(Self.draftTargetID)
-                            }
-                        }
-                        .padding(.horizontal, 2)
-                        .frame(
-                            width: max(layout.contentWidth + 4, trackWidth),
-                            height: ShellSidebarSpaceSliderLayout.trackHeight,
-                            alignment: .leading
-                        )
-                        .background {
-                            GeometryReader { contentProxy in
-                                Color.clear.preference(
-                                    key: ShellSidebarSpaceSliderScrollOffsetKey.self,
-                                    value: max(
-                                        0,
-                                        -contentProxy.frame(in: .named("spaceSliderTrack")).minX
-                                    )
-                                )
-                            }
-                        }
-                    }
-                    .frame(width: trackWidth, height: ShellSidebarSpaceSliderLayout.trackHeight)
-                    .clipShape(
-                        RoundedRectangle(
-                            cornerRadius: ShellSidebarSpaceSliderLayout.trackHeight * 0.5,
-                            style: .continuous
-                        )
-                    )
-                }
-                .frame(width: trackWidth, height: ShellSidebarSpaceSliderLayout.trackHeight)
-                .coordinateSpace(name: "spaceSliderTrack")
-                .padding(.horizontal, ShellSidebarMetrics.edgeInset)
-                .padding(.vertical, 2)
-                .contentShape(Rectangle())
-                .simultaneousGesture(dragScrubGesture(layout: layout))
-                .background {
-                    ShellSidebarSpaceSliderWheelMonitor(
-                        onScroll: { event, deltaX, deltaY in
-                            handleWheel(
-                                event: event,
-                                deltaX: deltaX,
-                                deltaY: deltaY,
-                                availableWidth: proxy.size.width
-                            )
-                        },
-                        onReset: resetWheelIntent,
-                        onContextMenuIntent: cancelScrubPreview
-                    )
-                }
-                .onPreferenceChange(ShellSidebarSpaceSliderScrollOffsetKey.self) { offset in
-                    trackScrollOffsetX = offset
-                }
-                .onChange(of: autoScrollSpaceID) { _, spaceID in
-                    scrollSpaceIntoView(spaceID, scrollProxy: scrollProxy)
-                }
-                .onAppear {
-                    scrollSpaceIntoView(autoScrollSpaceID, scrollProxy: scrollProxy, animated: false)
-                }
-            }
-            .offset(x: reduceMotion ? 0 : scrubState?.edgeResistanceOffset ?? 0)
-            .focusable()
-            .focused($isKeyboardFocused)
-            .focusEffectDisabled()
-            .onMoveCommand(perform: handleMoveCommand)
-            .onExitCommand {
-                cancelScrubPreview()
-            }
-            .onKeyPress(.return) {
-                commitKeyboardScrub()
-                return .handled
-            }
-            .onKeyPress(.space) {
-                commitKeyboardScrub()
-                return .handled
-            }
-        }
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .onChange(of: visibleSpaceIDs) { _, _ in
-            cancelInvalidScrubIfNeeded()
-        }
-        .onChange(of: resolvedDisplaySpaceID) { _, _ in
-            cancelScrubPreview()
-        }
-        .task {
-            spaceCreationProfileOptions.refresh()
-        }
-        .onChange(of: creationDraft != nil) { _, creating in
-            // Entering the creation form must cancel any pending scrub and
-            // invalidate the scheduled wheel-commit so it cannot fire and move
-            // the underlying Space behind the draft.
-            if creating {
-                cancelScrubPreview()
-                spaceCreationProfileOptions.refresh()
-            }
-        }
-    }
-
-    private var visibleSpaces: [ShellSpace] {
-        host.spaces
-    }
-
-    private func sliderLayout(availableWidth: CGFloat) -> ShellSidebarSpaceSliderLayout {
-        // While creating a Space, append one trailing draft target that is
-        // always the selected one; real-target hover/scrub focus is suppressed.
-        if creationDraft != nil {
-            return ShellSidebarSpaceSliderLayout.make(
-                spaceCount: visibleSpaces.count + 1,
-                selectedIndex: visibleSpaces.count,
-                hoveredIndex: nil,
-                scrubFocusIndex: nil,
-                availableWidth: availableWidth - (ShellSidebarMetrics.edgeInset * 2),
-                reduceMotion: reduceMotion
-            )
-        }
-
-        let selectedIndex = visibleSpaces.firstIndex { $0.spaceID == resolvedDisplaySpaceID }
-        let hoveredIndex = scrubState == nil
-            ? visibleSpaces.firstIndex { $0.spaceID == hoveredSpaceID }
-            : nil
-        let previewedIndex = visibleSpaces.firstIndex { $0.spaceID == scrubFocusSpaceID }
-        return ShellSidebarSpaceSliderLayout.make(
-            spaceCount: visibleSpaces.count,
-            selectedIndex: selectedIndex,
-            hoveredIndex: hoveredIndex,
-            scrubFocusIndex: previewedIndex,
-            availableWidth: availableWidth - (ShellSidebarMetrics.edgeInset * 2),
-            reduceMotion: reduceMotion
-        )
-    }
-
-    private func spaceControl(for space: ShellSpace, item: ShellSidebarSpaceSliderLayout.Item) -> some View {
-        Button {
-            // Selection is frozen on the draft target during Space creation.
-            guard creationDraft == nil else { return }
-            guard let targetIndex = ShellSidebarSpaceSliderClickSelection.targetIndex(
-                selectedIndex: selectedSpaceIndex,
-                clickedIndex: item.index,
-                spaceCount: visibleSpaces.count
-            ) else {
-                return
-            }
-            cancelScrubPreview()
-            host.select(spaceID: visibleSpaces[targetIndex].spaceID)
-        } label: {
-            ShellSidebarSpaceTrackTarget(
-                title: space.title,
-                icon: ShellSpacePresentationIcon.resolve(
-                    systemName: space.presentationIconSystemName,
-                    title: space.title
-                ),
-                mode: item.mode,
-                attention: strongestAttention(for: space),
-                isSelected: item.isSelected,
-                isFocused: item.isFocused,
-                isHovered: hoveredSpaceID == space.spaceID
-            )
-        }
-        .buttonStyle(.plain)
-        .frame(width: item.width, height: ShellSidebarSpaceSliderLayout.itemHeight)
-        .contentShape(Rectangle())
-        .contextMenu {
-            spaceContextMenu(for: space)
-        }
-        .help(space.title)
-        .accessibilityLabel(spaceAccessibilityLabel(for: space, isSelected: item.isSelected))
-        .accessibilityAddTraits(item.isSelected ? .isSelected : [])
-        .onHover { isHovering in
-            hoveredSpaceID = isHovering ? space.spaceID : nil
-        }
-    }
-
-    /// Display-only trailing target for the in-progress draft Space. Mirrors the
-    /// real `spaceControl` track target but attaches no tap/scrub/context
-    /// affordances — selection cannot move to it and it commits nothing.
-    private func draftControl(
-        for draft: ShellSpaceSliderDraft,
-        item: ShellSidebarSpaceSliderLayout.Item
-    ) -> some View {
-        ShellSidebarSpaceTrackTarget(
-            title: draft.name.trimmingCharacters(in: .whitespacesAndNewlines),
-            icon: ShellSpacePresentationIcon.resolve(
-                systemName: draft.iconSystemName,
-                title: draft.name
-            ),
-            mode: item.mode,
-            attention: .idle,
-            isSelected: item.isSelected,
-            isFocused: item.isFocused,
-            isHovered: false
-        )
-        .frame(width: item.width, height: ShellSidebarSpaceSliderLayout.itemHeight)
-        .contentShape(Rectangle())
-        .allowsHitTesting(false)
-        .accessibilityLabel("New Space draft")
-        .accessibilityAddTraits(.isSelected)
-    }
-
-    @ViewBuilder
-    private func spaceContextMenu(for space: ShellSpace) -> some View {
-        Menu("Terminal Profile") {
-            Button {
-                cancelScrubPreview()
-                _ = host.setTerminalProfile(nil, forSpaceID: space.spaceID)
-            } label: {
-                Label("Login shell", systemImage: space.terminalProfileID == nil ? "checkmark" : "terminal")
-            }
-
-            ForEach(spaceCreationProfileOptions.options) { option in
-                Button {
-                    cancelScrubPreview()
-                    _ = host.setTerminalProfile(option.id, forSpaceID: space.spaceID)
-                } label: {
-                    Label(
-                        option.name,
-                        systemImage: option.id == space.terminalProfileID
-                            ? "checkmark"
-                            : option.systemName
-                    )
-                }
-                .disabled(!option.isEnabled)
-                .help(option.guidance ?? option.name)
-            }
-        }
-
-        Menu("Space Icon") {
-            Button {
-                cancelScrubPreview()
-                _ = host.setPresentationIcon(nil, forSpaceID: space.spaceID)
-            } label: {
-                Label(
-                    "Default (Initial)",
-                    systemImage: space.presentationIconSystemName == nil ? "checkmark" : "textformat"
-                )
-            }
-
-            Divider()
-
-            ForEach(ShellSpaceIconCatalog.curatedSymbols, id: \.self) { symbol in
-                Button {
-                    cancelScrubPreview()
-                    _ = host.setPresentationIcon(symbol, forSpaceID: space.spaceID)
-                } label: {
-                    Label(
-                        iconMenuTitle(symbol),
-                        systemImage: space.presentationIconSystemName == symbol ? "checkmark" : symbol
-                    )
-                }
-            }
-        }
-    }
-
-    private var resolvedDisplaySpaceID: String? {
-        displaySpaceID ?? host.selectedSpace?.spaceID
-    }
-
-    private var selectedSpaceIndex: Int? {
-        visibleSpaces.firstIndex { $0.spaceID == resolvedDisplaySpaceID }
-    }
-
-    private var scrubFocusSpaceID: String? {
-        if let scrubState,
-           visibleSpaces.indices.contains(scrubState.focusIndex)
-        {
-            return visibleSpaces[scrubState.focusIndex].spaceID
-        }
-        return previewedSpaceID
-    }
-
-    private var autoScrollSpaceID: String? {
-        // While creating, keep the trailing draft target on-screen so its live
-        // identity preview stays visible even when the slider is scrollable.
-        if creationDraft != nil { return Self.draftTargetID }
-        return scrubFocusSpaceID ?? resolvedDisplaySpaceID
-    }
-
-    private var visibleSpaceIDs: [String] {
-        visibleSpaces.map(\.spaceID)
-    }
-
-    private func dragScrubGesture(layout: ShellSidebarSpaceSliderLayout) -> some Gesture {
-        DragGesture(
-            minimumDistance: ShellSidebarSpaceSliderScrubState.dragThreshold,
-            coordinateSpace: .local
-        )
-        .onChanged { value in
-            updateDragScrub(value, layout: layout)
-        }
-        .onEnded { value in
-            updateDragScrub(value, layout: layout)
-            commitScrubSelection()
-        }
-    }
-
-    private func updateDragScrub(
-        _ value: DragGesture.Value,
-        layout: ShellSidebarSpaceSliderLayout
-    ) {
-        guard var nextState = scrubStateForUpdating(source: .drag) else { return }
-        nextState.updateDrag(
-            locationX: value.location.x - ShellSidebarMetrics.edgeInset + trackScrollOffsetX,
-            translationX: value.translation.width,
-            layout: layout
-        )
-        scrubState = nextState
-    }
-
-    private func handleWheel(
-        event: ShellSidebarSpaceSliderWheelEvent,
-        deltaX: CGFloat,
-        deltaY: CGFloat,
-        availableWidth: CGFloat
-    ) -> Bool {
-        var nextIntent = wheelIntentState
-        let route = nextIntent.route(deltaX: deltaX, deltaY: deltaY)
-        wheelIntentState = nextIntent
-
-        switch route {
-        case .passThrough:
-            guard ShellSidebarSpaceSliderWheelForwarding.shouldForwardPassThroughToTabList(
-                deltaX: deltaX,
-                deltaY: deltaY
-            ) else {
-                return false
-            }
-            return onForwardVerticalWheel(event)
-        case .scrub(let routedDeltaX):
-            updateWheelScrub(deltaX: routedDeltaX, availableWidth: availableWidth)
-            scheduleWheelCommit()
-            return true
-        }
-    }
-
-    private func updateWheelScrub(deltaX: CGFloat, availableWidth: CGFloat) {
-        guard var nextState = scrubStateForUpdating(source: .wheel) else { return }
-        let itemSpan = max(
-            (availableWidth - ShellSidebarMetrics.edgeInset * 2)
-                / CGFloat(max(visibleSpaces.count, 1)),
-            ShellSidebarSpaceSliderScrubState.wheelStepWidth
-        )
-        nextState.updateWheel(
-            deltaX: deltaX,
-            itemSpan: itemSpan,
-            spaceCount: visibleSpaces.count
-        )
-        scrubState = nextState
-    }
-
-    private func handleMoveCommand(_ direction: MoveCommandDirection) {
-        let delta: Int
-        switch direction {
-        case .left:
-            delta = -1
-        case .right:
-            delta = 1
-        default:
-            return
-        }
-
-        guard var nextState = scrubStateForUpdating(source: .keyboard) else { return }
-        nextState.moveFocus(by: delta, spaceCount: visibleSpaces.count)
-        scrubState = nextState
-    }
-
-    private func scrubStateForUpdating(
-        source: ShellSidebarSpaceSliderScrubSource
-    ) -> ShellSidebarSpaceSliderScrubState? {
-        // No scrub/keyboard/wheel selection while a draft target owns selection.
-        guard creationDraft == nil else { return nil }
-        if let scrubState, scrubState.source == source {
-            return scrubState
-        }
-        return ShellSidebarSpaceSliderScrubState(
-            source: source,
-            selectedIndex: selectedSpaceIndex,
-            spaceCount: visibleSpaces.count
-        )
-    }
-
-    private func scheduleWheelCommit() {
-        wheelCommitToken += 1
-        let token = wheelCommitToken
-        DispatchQueue.main.asyncAfter(
-            deadline: .now() + ShellSidebarSpaceSliderScrubState.wheelCommitDelay
-        ) {
-            guard wheelCommitToken == token,
-                  scrubState?.source == .wheel
-            else {
-                return
-            }
-            commitScrubSelection()
-            resetWheelIntent()
-        }
-    }
-
-    private func scrollSpaceIntoView(
-        _ spaceID: String?,
-        scrollProxy: ScrollViewProxy,
-        animated: Bool = true
-    ) {
-        guard let spaceID else { return }
-        let action = {
-            scrollProxy.scrollTo(spaceID, anchor: .center)
-        }
-        if animated, !reduceMotion {
-            withAnimation(.easeOut(duration: 0.14), action)
-        } else {
-            action()
-        }
-    }
-
-    private func commitKeyboardScrub() {
-        guard scrubState?.source == .keyboard else { return }
-        commitScrubSelection()
-    }
-
-    private func commitScrubSelection() {
-        // Chokepoint for every scrub commit (wheel timer, drag-end, keyboard).
-        // While a draft owns selection, a stale in-flight commit must not move
-        // the underlying Space — drop it and clear any lingering scrub state.
-        guard creationDraft == nil else {
-            cancelScrubPreview()
-            return
-        }
-        guard let scrubState,
-              visibleSpaces.indices.contains(scrubState.commitIndex)
-        else {
-            cancelScrubPreview()
-            return
-        }
-
-        let targetSpace = visibleSpaces[scrubState.commitIndex]
-        cancelScrubPreview()
-        if targetSpace.spaceID != resolvedDisplaySpaceID {
-            host.select(spaceID: targetSpace.spaceID)
-        }
-    }
-
-    private func cancelScrubPreview() {
-        scrubState = nil
-        wheelCommitToken += 1
-        resetWheelIntent()
-    }
-
-    private func resetWheelIntent() {
-        wheelIntentState.reset()
-    }
-
-    private func cancelInvalidScrubIfNeeded() {
-        guard let scrubState else { return }
-        guard visibleSpaces.indices.contains(scrubState.focusIndex),
-              visibleSpaces.indices.contains(scrubState.sourceIndex)
-        else {
-            cancelScrubPreview()
-            return
-        }
-    }
-
-    /// Readable label for a curated Space-icon SF Symbol. Falls back to a
-    /// title-cased transform of the raw symbol name for any unmapped symbol.
-    private func iconMenuTitle(_ symbol: String) -> String {
-        switch symbol {
-        case "terminal": return "Terminal"
-        case "chevron.left.forwardslash.chevron.right": return "Code"
-        case "hammer": return "Build"
-        case "wrench.and.screwdriver": return "Tools"
-        case "ant": return "Debug"
-        case "flask": return "Experiment"
-        case "cube.box": return "Package"
-        case "shippingbox": return "Release"
-        case "server.rack": return "Server"
-        case "externaldrive": return "Storage"
-        case "doc.text": return "Document"
-        case "book": return "Docs"
-        case "paintbrush": return "Design"
-        case "paintpalette": return "Palette"
-        case "globe": return "Web"
-        case "network": return "Network"
-        case "lock": return "Secure"
-        case "key": return "Keys"
-        case "leaf": return "Nature"
-        case "bolt": return "Power"
-        case "sparkles": return "Sparkle"
-        case "star": return "Star"
-        case "flag": return "Flag"
-        case "folder": return "Folder"
-        default:
-            return symbol
-                .split(separator: ".")
-                .map(\.capitalized)
-                .joined(separator: " ")
-        }
-    }
-
-    private func strongestAttention(for space: ShellSpace) -> ShellAttentionState {
-        host.shellState.panes
-            .filter { $0.spaceID == space.spaceID }
-            .map { shellEffectiveAttention(for: $0, now: activityFreshnessNow) }
-            .filter(\.requiresUserAction)
-            .max(by: { attentionRank(for: $0) < attentionRank(for: $1) })
-            ?? .idle
-    }
-
-    private func attentionRank(for attention: ShellAttentionState) -> Int {
-        switch attention {
-        case .awaitingUser:
-            return 3
-        case .notable:
-            return 2
-        case .active:
-            return 1
-        case .idle:
-            return 0
-        }
-    }
-
-    private func spaceAccessibilityLabel(for space: ShellSpace, isSelected: Bool) -> String {
-        var parts = [space.title, space.tabs.count == 1 ? "1 tab" : "\(space.tabs.count) tabs"]
-        if isSelected {
-            parts.append("selected")
-        }
-        if previewedSpaceID == space.spaceID {
-            parts.append("preview")
-        }
-        if strongestAttention(for: space) != .idle {
-            parts.append("needs attention")
-        }
-        return parts.joined(separator: ", ")
-    }
-}
-
-private struct ShellSidebarSpaceTrackTarget: View {
-    let title: String
-    let icon: ShellSpacePresentationIcon.Resolved
-    let mode: ShellSidebarSpaceSliderLayout.DisplayMode
-    let attention: ShellAttentionState
-    let isSelected: Bool
-    let isFocused: Bool
-    let isHovered: Bool
-
-    var body: some View {
-        HStack(spacing: textSpacing) {
-            spaceIcon
-                .frame(width: 15, height: 15)
-
-            if mode != .iconOnly {
-                Text(title)
-                    .font(font)
-                    .foregroundStyle(textForeground)
-                    .lineLimit(1)
-                    .truncationMode(.tail)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-            }
-        }
-        .padding(.horizontal, horizontalPadding)
-        .frame(
-            maxWidth: .infinity,
-            minHeight: ShellSidebarSpaceSliderLayout.itemHeight,
-            alignment: .center
-        )
-        .background {
-            if isSelected {
-                ShellLiquidGlassSurface(
-                    shape: Capsule(),
-                    tint: ShellPalette.sidebarSelection,
-                    tintOpacity: 0.30,
-                    strokeOpacity: 0.24,
-                    usesSystemGlassInLightMode: true
-                )
-            } else if isFocused || isHovered {
-                Capsule()
-                    .strokeBorder(focusStroke, lineWidth: 0.7)
-            }
-        }
-    }
-
-    @ViewBuilder
-    private var spaceIcon: some View {
-        switch icon {
-        case .symbol(let name), .fallbackSymbol(let name):
-            Image(systemName: name)
-                .font(.system(size: iconSize, weight: .semibold))
-                .symbolRenderingMode(.hierarchical)
-                .foregroundStyle(iconForeground)
-                .accessibilityHidden(true)
-        case .monogram(let text):
-            Text(text)
-                .font(ShellType.pro(iconSize, weight: .semibold))
-                .foregroundStyle(iconForeground)
-                .lineLimit(1)
-                .minimumScaleFactor(0.7)
-                .accessibilityHidden(true)
-        }
-    }
-
-    private var font: Font {
-        switch mode {
-        case .fullTitle:
-            return .system(size: 12, weight: isSelected ? .semibold : .medium)
-        case .truncatedTitle:
-            return .system(size: 11.6, weight: isSelected || isFocused ? .semibold : .medium)
-        case .iconOnly:
-            return .system(size: 11.4, weight: .medium)
-        }
-    }
-
-    private var horizontalPadding: CGFloat {
-        switch mode {
-        case .fullTitle:
-            return 8
-        case .truncatedTitle:
-            return 6
-        case .iconOnly:
-            return 0
-        }
-    }
-
-    private var textSpacing: CGFloat {
-        switch mode {
-        case .fullTitle:
-            return 5
-        case .truncatedTitle:
-            return 4
-        case .iconOnly:
-            return 0
-        }
-    }
-
-    private var iconSize: CGFloat {
-        switch mode {
-        case .fullTitle, .truncatedTitle:
-            return 11.4
-        case .iconOnly:
-            return 12.4
-        }
-    }
-
-    private var textForeground: Color {
-        if isSelected {
-            return ShellPalette.sidebarInk.opacity(0.92)
-        }
-        if isFocused || isHovered {
-            return ShellPalette.sidebarInk.opacity(0.82)
-        }
-        return ShellPalette.sidebarMutedInk.opacity(0.72)
-    }
-
-    private var iconForeground: Color {
-        if attention.requiresUserAction {
-            return ShellSignal.action.opacity(isSelected ? 0.94 : 0.86)
-        }
-        if isSelected {
-            return ShellPalette.sidebarInk.opacity(0.92)
-        }
-        if isFocused {
-            return ShellPalette.accent.opacity(0.82)
-        }
-        if isHovered {
-            return ShellPalette.sidebarInk.opacity(0.78)
-        }
-        return ShellPalette.sidebarMutedInk.opacity(0.62)
-    }
-
-    private var focusStroke: Color {
-        if isFocused {
-            return ShellPalette.accent.opacity(0.34)
-        }
-        return ShellPalette.sidebarInk.opacity(0.14)
-    }
-}
-
-private enum ShellSidebarRowVisualState: Equatable {
-    case normal
-    case hover
-    case selected
-
-    var cornerRadius: CGFloat {
-        switch self {
-        case .normal:
-            return ShellRadii.row
-        case .hover:
-            return ShellRadii.control
-        case .selected:
-            return ShellRadii.row
-        }
-    }
-
-    var fill: Color? {
-        switch self {
-        case .normal:
-            return nil
-        case .hover:
-            return ShellPalette.sidebarRowHover
-        case .selected:
-            return ShellPalette.sidebarRowSelected
-        }
-    }
-
-    var stroke: Color {
-        switch self {
-        case .normal:
-            return .clear
-        case .hover:
-            return ShellPalette.line.opacity(0.08)
-        case .selected:
-            return ShellPalette.line.opacity(0.12)
-        }
-    }
-
-    var shadow: ShellShadowStyle {
-        switch self {
-        case .normal, .hover:
-            return ShellShadows.none
-        case .selected:
-            return ShellShadows.navigationSelection
-        }
-    }
-}
-
-private struct ShellSidebarRowBackground: View {
-    @Environment(\.colorScheme) private var colorScheme
-    let state: ShellSidebarRowVisualState
-
-    var body: some View {
-        if let fill = state.fill {
-            let shape = RoundedRectangle(cornerRadius: state.cornerRadius, style: .continuous)
-            shape
-                .fill(fill)
-                .overlay {
-                    shape.stroke(state.stroke, lineWidth: 0.5)
-                }
-                .overlay {
-                    if colorScheme == .light && state == .selected {
-                        shape
-                            .stroke(Color.white.opacity(0.34), lineWidth: 0.55)
-                            .mask {
-                                shape.fill(
-                                    LinearGradient(
-                                        colors: [
-                                            Color.white,
-                                            Color.white.opacity(0),
-                                        ],
-                                        startPoint: .top,
-                                        endPoint: .bottom
-                                    )
-                                )
-                            }
-                    }
-                }
-                .shellShadow(state.shadow)
-        }
-    }
-}
-
-private enum ShellSidebarTypography {
-    static let titleSize: CGFloat = 14
-    static let markerSize: CGFloat = 9
-    static let closeSize: CGFloat = 9.5
-
-    static func titleWeight(isSelected: Bool) -> Font.Weight {
-        isSelected ? .medium : .regular
-    }
-
-    static let secondaryWeight: Font.Weight = .regular
-    static let secondaryEmphasisWeight: Font.Weight = .medium
-    static let iconWeight: Font.Weight = .medium
-    static let markerWeight: Font.Weight = .semibold
-}
-
-private enum ShellSidebarRowMetrics {
-    static let height: CGFloat = 36
-    static let horizontalInset: CGFloat = 8
-    static let leadingSlot: CGFloat = 24
-    static let trailingSlot: CGFloat = 20
-    static let dragMidpoint: CGFloat = height / 2
-}
-
-private enum ShellSidebarTabListMetrics {
-    static let itemSpacing: CGFloat = 6
-    static let sliderToListLift: CGFloat = 12
-}
-
-private enum ShellSidebarTabDragState {
-    static let dragThreshold: CGFloat = 7
-}
-
-private enum ShellSidebarTabControlMetrics {
-    static let hitHeight: CGFloat = 16
-    static let horizontalInset: CGFloat = ShellSidebarRowMetrics.horizontalInset
-}
-
-private struct ShellSidebarTabControlRow: View {
-    @State private var isControlHovered = false
-    let showsDivider: Bool
-    let showsClear: Bool
-    let isClearEnabled: Bool
-    let clearAction: () -> Void
-
-    var body: some View {
-        HStack(alignment: .center, spacing: 8) {
-            if showsDivider {
-                Rectangle()
-                    .fill(ShellPalette.sidebarDivider)
-                    .frame(maxWidth: .infinity, minHeight: 1, maxHeight: 1)
-            }
-
-            if showsClear && isControlHovered {
-                Button(action: clearAction) {
-                    Label("Clear", systemImage: "arrow.down.to.line.compact")
-                        .labelStyle(.titleAndIcon)
-                        .font(.system(size: 11, weight: .semibold))
-                }
-                .buttonStyle(.plain)
-                .disabled(!isClearEnabled)
-                .foregroundStyle(ShellPalette.sidebarMutedInk.opacity(isClearEnabled ? 0.72 : 0.44))
-                .help("Clear inactive tabs")
-                .fixedSize()
-            }
-        }
-        .padding(.horizontal, ShellSidebarTabControlMetrics.horizontalInset)
-        .frame(
-            maxWidth: .infinity,
-            minHeight: ShellSidebarTabControlMetrics.hitHeight,
-            maxHeight: ShellSidebarTabControlMetrics.hitHeight,
-            alignment: .center
-        )
-        .contentShape(Rectangle())
-        .onHover { isControlHovered = $0 }
-        .animation(.easeInOut(duration: 0.12), value: isControlHovered)
-    }
-}
-
-private struct ShellTabSidebarRow: View {
-    @Environment(\.accessibilityReduceMotion) private var reduceMotion
-    @FocusState private var isKeyboardFocused: Bool
-    @State private var isCloseHovered = false
-    let title: String
-    let subtitle: String?
-    let isActivitySubtitle: Bool
-    let secondaryIsMachineFact: Bool
-    let progress: TerminalActivityProgress?
-    let stateAccessory: ShellSidebarTabStateAccessory?
-    let attention: ShellAttentionState?
-    let showsAlanMarker: Bool
-    let paneSummary: ShellTabPaneSummary?
-    let isPinned: Bool
-    let isSelected: Bool
-    let isHovered: Bool
-    let showsCloseAffordance: Bool
-    let onFocusSplitPane: (String) -> Void
-    let onFocusNextSplitPane: (ShellTabPaneSummary) -> Void
-    let onClose: () -> Void
-
-    var body: some View {
-        HStack(alignment: .center, spacing: 8) {
-            leadingSlot
-                .frame(width: ShellSidebarRowMetrics.leadingSlot, height: ShellSidebarRowMetrics.leadingSlot, alignment: .center)
-
-            VStack(alignment: .leading, spacing: subtitle == nil ? 0 : 1) {
-                HStack(spacing: 6) {
-                    Text(title)
-                        .font(
-                            .system(
-                                size: ShellSidebarTypography.titleSize,
-                                weight: ShellSidebarTypography.titleWeight(isSelected: isSelected)
-                            )
-                        )
-                        .foregroundStyle(titleForeground)
-                        .lineLimit(1)
-                        .truncationMode(.middle)
-
-                    if showsAlanMarker {
-                        Image(systemName: "sparkles")
-                            .font(
-                                .system(
-                                    size: ShellSidebarTypography.markerSize,
-                                    weight: ShellSidebarTypography.markerWeight
-                                )
-                            )
-                            .foregroundStyle(ShellPalette.accent)
-                    }
-
-                }
-
-                if let subtitle {
-                    subtitleText(subtitle)
-                        .foregroundStyle(subtitleForeground)
-                        .lineLimit(1)
-                        .truncationMode(.middle)
-                }
-
-                if let progress {
-                    ShellSidebarActivityProgressRail(progress: progress, isSelected: isSelected)
-                }
-            }
-            .frame(maxWidth: .infinity, alignment: .leading)
-
-            closeButtonSlot
-        }
-        .padding(.horizontal, ShellSidebarRowMetrics.horizontalInset)
-        .frame(minHeight: ShellSidebarRowMetrics.height)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background(
-            ShellSidebarRowBackground(state: visualState)
-        )
-        .contentShape(RoundedRectangle(cornerRadius: visualState.cornerRadius, style: .continuous))
-        .animation(reduceMotion ? nil : .easeOut(duration: 0.14), value: visualState)
-        .animation(reduceMotion ? nil : .easeOut(duration: 0.12), value: showsCloseButton)
-        .focusable()
-        .focused($isKeyboardFocused)
-        .focusEffectDisabled()
-        .accessibilityLabel(accessibilityLabel)
-        .help("Select tab")
-    }
-
-    private var visualState: ShellSidebarRowVisualState {
-        if isSelected {
-            return .selected
-        }
-
-        if isHovered || isKeyboardFocused {
-            return .hover
-        }
-
-        return .normal
-    }
-
-    private var isInteractionActive: Bool {
-        isHovered || showsCloseAffordance || isKeyboardFocused
-    }
-
-    private var showsCloseButton: Bool {
-        isSelected || isInteractionActive
-    }
-
-    @ViewBuilder
-    private var closeButtonSlot: some View {
-        ZStack {
-            if showsCloseButton {
-                Button(action: onClose) {
-                    Image(systemName: "xmark")
-                        .font(
-                            .system(
-                                size: ShellSidebarTypography.closeSize,
-                                weight: ShellSidebarTypography.markerWeight
-                            )
-                        )
-                        .foregroundStyle(closeForeground)
-                        .frame(width: ShellSidebarRowMetrics.trailingSlot, height: ShellSidebarRowMetrics.trailingSlot)
-                        .contentShape(Circle())
-                        .background {
-                            if isCloseHovered || isKeyboardFocused {
-                                Circle()
-                                    .fill(ShellPalette.sidebarInk.opacity(isSelected ? 0.05 : 0.035))
-                            }
-                        }
-                }
-                .buttonStyle(.plain)
-                .help("Close tab")
-                .accessibilityLabel("Close tab")
-                .accessibilityHidden(!showsCloseButton)
-                .onHover { isHovering in
-                    isCloseHovered = isHovering
-                }
-            } else if let stateAccessory {
-                Image(systemName: stateAccessory.systemImageName)
-                    .font(.system(size: 10.5, weight: .semibold))
-                    .foregroundStyle(ShellPalette.sidebarMutedInk.opacity(0.70))
-                    .frame(width: ShellSidebarRowMetrics.trailingSlot, height: ShellSidebarRowMetrics.trailingSlot)
-                    .help(stateAccessory.accessibilityLabel)
-                    .accessibilityLabel(stateAccessory.accessibilityLabel)
-            } else {
-                Color.clear
-                    .frame(width: ShellSidebarRowMetrics.trailingSlot, height: ShellSidebarRowMetrics.trailingSlot)
-                    .accessibilityHidden(true)
-            }
-        }
-    }
-
-    @ViewBuilder
-    private var leadingSlot: some View {
-        if let paneSummary {
-            ShellPaneTopologyIndicator(
-                summary: paneSummary,
-                isSelected: isSelected,
-                onFocusSplitPane: onFocusSplitPane,
-                onFocusNextSplitPane: onFocusNextSplitPane
-            )
-        } else {
-            ShellPaneTopologyIndicator.placeholder(isSelected: isSelected)
-        }
-    }
-
-    private func subtitleText(_ subtitle: String) -> Text {
-        let parts = subtitle.components(separatedBy: " · ")
-        guard isActivitySubtitle, !parts.isEmpty else {
-            // Only machine facts (cwd/branch/process context) render in the mono
-            // accent track; human-language status summaries and content-type
-            // hints stay in SF Pro. See docs/design/design-language.md,
-            // principle 4.
-            if secondaryIsMachineFact {
-                return Text(subtitle)
-                    .font(ShellType.mono(ShellType.monoCaption, weight: ShellSidebarTypography.secondaryWeight))
-            }
-            return Text(subtitle)
-                .font(ShellType.pro(ShellType.caption, weight: ShellSidebarTypography.secondaryWeight))
-        }
-
-        let emphasizedIndex = emphasizedSubtitleIndex(for: parts)
-        var attributedSubtitle = AttributedString()
-        for element in parts.enumerated() {
-            let (index, part) = element
-            let prefix = index == 0 ? "" : " · "
-            let weight = index == emphasizedIndex
-                ? ShellSidebarTypography.secondaryEmphasisWeight
-                : ShellSidebarTypography.secondaryWeight
-            var fragment = AttributedString(prefix + part)
-            fragment.font = ShellType.pro(ShellType.caption, weight: weight)
-            attributedSubtitle += fragment
-        }
-        return Text(attributedSubtitle)
-    }
-
-    private func emphasizedSubtitleIndex(for parts: [String]) -> Int {
-        if parts.count >= 3,
-           parts[0].hasPrefix("Pane ")
-        {
-            return 1
-        }
-        return 0
-    }
-
-    private var titleForeground: Color {
-        isSelected ? ShellPalette.sidebarInk : ShellPalette.sidebarInk.opacity(0.88)
-    }
-
-    private var subtitleForeground: Color {
-        isSelected ? ShellPalette.sidebarMutedInk.opacity(0.90) : ShellPalette.sidebarMutedInk.opacity(0.68)
-    }
-
-    private var closeForeground: Color {
-        if isCloseHovered {
-            return ShellPalette.sidebarInk.opacity(isSelected ? 0.68 : 0.76)
-        }
-        return isSelected ? ShellPalette.sidebarInk.opacity(0.46) : ShellPalette.sidebarMutedInk.opacity(0.62)
-    }
-
-    private var accessibilityLabel: String {
-        var parts = [title]
-        if let subtitle {
-            parts.append(subtitle)
-        }
-        if isSelected {
-            parts.append("selected")
-        }
-        if attention != nil {
-            parts.append("needs attention")
-        }
-        if let paneSummary {
-            parts.append(paneSummary.paneCount == 1 ? "1 pane" : "\(paneSummary.paneCount) panes")
-        }
-        if isPinned {
-            parts.append("pinned")
-        }
-        return parts.joined(separator: ", ")
-    }
-}
-
-private struct ShellSidebarActivityProgressRail: View {
-    let progress: TerminalActivityProgress
-    let isSelected: Bool
-
-    var body: some View {
-        GeometryReader { proxy in
-            ZStack(alignment: .leading) {
-                Capsule()
-                    .fill(ShellPalette.sidebarMutedInk.opacity(isSelected ? 0.16 : 0.12))
-
-                Capsule()
-                    .fill(fillColor)
-                    .frame(width: fillWidth(in: proxy.size.width))
-            }
-        }
-        .frame(height: 2)
-        .accessibilityHidden(true)
-    }
-
-    private var fillColor: Color {
-        switch progress.kind {
-        case .failed:
-            return ShellSignal.action.opacity(isSelected ? 0.86 : 0.72)
-        case .paused:
-            return ShellPalette.sidebarMutedInk.opacity(isSelected ? 0.62 : 0.48)
-        case .percent, .indeterminate:
-            return ShellPalette.accent.opacity(isSelected ? 0.82 : 0.68)
-        }
-    }
-
-    private func fillWidth(in width: CGFloat) -> CGFloat {
-        switch progress.kind {
-        case .percent:
-            return width * CGFloat(progress.percent ?? 0) / 100
-        case .indeterminate:
-            return max(width * 0.36, 18)
-        case .paused, .failed:
-            return width
-        }
-    }
-}
-
-private struct ShellPaneTopologyIndicator: View {
-    let summary: ShellTabPaneSummary
-    let isSelected: Bool
-    let onFocusSplitPane: (String) -> Void
-    let onFocusNextSplitPane: (ShellTabPaneSummary) -> Void
-
-    @ViewBuilder
-    var body: some View {
-        switch summary.topology.kind {
-        case .single:
-            singlePaneIndicator
-        case .columns(let count):
-            columnsIndicator(paneIDs: Array(summary.paneIDs.prefix(count)))
-        case .rows(let count):
-            rowsIndicator(paneIDs: Array(summary.paneIDs.prefix(count)))
-        case .mainLeftWithRightStack:
-            mainLeftWithRightStackIndicator
-        case .mainRightWithLeftStack:
-            mainRightWithLeftStackIndicator
-        case .mainTopWithBottomSplit:
-            mainTopWithBottomSplitIndicator
-        case .mainBottomWithTopSplit:
-            mainBottomWithTopSplitIndicator
-        case .grid2x2(let rootDirection):
-            gridIndicator(rootDirection: rootDirection)
-        case .complex:
-            complexButton
-        }
-    }
-
-    private var singlePaneIndicator: some View {
-        indicatorFrame {
-            RoundedRectangle(cornerRadius: ShellRadii.micro, style: .continuous)
-                .fill(singlePaneFill)
-                .frame(maxWidth: .infinity, maxHeight: .infinity)
-        }
-        .accessibilityLabel("Single pane")
-    }
-
-    private func columnsIndicator(paneIDs: [String]) -> some View {
-        indicatorFrame {
-            HStack(spacing: 2) {
-                ForEach(paneIDs, id: \.self) { paneID in
-                    segmentButton(paneID: paneID)
-                }
-            }
-        }
-        .help("Focus split pane")
-        .accessibilityLabel(splitAccessibilityLabel)
-    }
-
-    private func rowsIndicator(paneIDs: [String]) -> some View {
-        indicatorFrame {
-            VStack(spacing: 2) {
-                ForEach(paneIDs, id: \.self) { paneID in
-                    segmentButton(paneID: paneID)
-                }
-            }
-        }
-        .help("Focus split pane")
-        .accessibilityLabel(splitAccessibilityLabel)
-    }
-
-    @ViewBuilder
-    private var mainLeftWithRightStackIndicator: some View {
-        let paneIDs = Array(summary.paneIDs.prefix(3))
-        if paneIDs.count == 3 {
-            indicatorFrame {
-                HStack(spacing: 2) {
-                    segmentButton(paneID: paneIDs[0])
-                    VStack(spacing: 2) {
-                        segmentButton(paneID: paneIDs[1])
-                        segmentButton(paneID: paneIDs[2])
-                    }
-                }
-            }
-            .help("Focus split pane")
-            .accessibilityLabel(splitAccessibilityLabel)
-        } else {
-            complexButton
-        }
-    }
-
-    @ViewBuilder
-    private var mainRightWithLeftStackIndicator: some View {
-        let paneIDs = Array(summary.paneIDs.prefix(3))
-        if paneIDs.count == 3 {
-            indicatorFrame {
-                HStack(spacing: 2) {
-                    VStack(spacing: 2) {
-                        segmentButton(paneID: paneIDs[0])
-                        segmentButton(paneID: paneIDs[1])
-                    }
-                    segmentButton(paneID: paneIDs[2])
-                }
-            }
-            .help("Focus split pane")
-            .accessibilityLabel(splitAccessibilityLabel)
-        } else {
-            complexButton
-        }
-    }
-
-    @ViewBuilder
-    private var mainTopWithBottomSplitIndicator: some View {
-        let paneIDs = Array(summary.paneIDs.prefix(3))
-        if paneIDs.count == 3 {
-            indicatorFrame {
-                VStack(spacing: 2) {
-                    segmentButton(paneID: paneIDs[0])
-                    HStack(spacing: 2) {
-                        segmentButton(paneID: paneIDs[1])
-                        segmentButton(paneID: paneIDs[2])
-                    }
-                }
-            }
-            .help("Focus split pane")
-            .accessibilityLabel(splitAccessibilityLabel)
-        } else {
-            complexButton
-        }
-    }
-
-    @ViewBuilder
-    private var mainBottomWithTopSplitIndicator: some View {
-        let paneIDs = Array(summary.paneIDs.prefix(3))
-        if paneIDs.count == 3 {
-            indicatorFrame {
-                VStack(spacing: 2) {
-                    HStack(spacing: 2) {
-                        segmentButton(paneID: paneIDs[0])
-                        segmentButton(paneID: paneIDs[1])
-                    }
-                    segmentButton(paneID: paneIDs[2])
-                }
-            }
-            .help("Focus split pane")
-            .accessibilityLabel(splitAccessibilityLabel)
-        } else {
-            complexButton
-        }
-    }
-
-    @ViewBuilder
-    private func gridIndicator(rootDirection: ShellSplitDirection) -> some View {
-        let paneIDs = Array(summary.paneIDs.prefix(4))
-        if paneIDs.count == 4 {
-            indicatorFrame {
-                if rootDirection == .vertical {
-                    HStack(spacing: 2) {
-                        VStack(spacing: 2) {
-                            segmentButton(paneID: paneIDs[0])
-                            segmentButton(paneID: paneIDs[1])
-                        }
-                        VStack(spacing: 2) {
-                            segmentButton(paneID: paneIDs[2])
-                            segmentButton(paneID: paneIDs[3])
-                        }
-                    }
-                } else {
-                    VStack(spacing: 2) {
-                        HStack(spacing: 2) {
-                            segmentButton(paneID: paneIDs[0])
-                            segmentButton(paneID: paneIDs[1])
-                        }
-                        HStack(spacing: 2) {
-                            segmentButton(paneID: paneIDs[2])
-                            segmentButton(paneID: paneIDs[3])
-                        }
-                    }
-                }
-            }
-            .help("Focus split pane")
-            .accessibilityLabel(splitAccessibilityLabel)
-        } else {
-            complexButton
-        }
-    }
-
-    private func indicatorFrame<Content: View>(@ViewBuilder content: () -> Content) -> some View {
-        let shape = RoundedRectangle(cornerRadius: ShellRadii.badge, style: .continuous)
-
-        return content()
-            .padding(3)
-            .frame(width: 22, height: 18)
-            .background {
-                shape
-                    .fill(containerFill)
-                    .overlay {
-                        shape.stroke(ShellPalette.line.opacity(isSelected ? 0.20 : 0.15), lineWidth: 0.5)
-                    }
-            }
-    }
-
-    static func placeholder(isSelected: Bool) -> some View {
-        let shape = RoundedRectangle(cornerRadius: ShellRadii.badge, style: .continuous)
-        return shape
-            .fill(ShellPalette.sidebarMutedInk.opacity(isSelected ? 0.20 : 0.13))
-            .frame(width: 22, height: 18)
-            .overlay {
-                shape.stroke(ShellPalette.line.opacity(isSelected ? 0.20 : 0.15), lineWidth: 0.5)
-            }
-            .accessibilityLabel("Pane")
-    }
-
-    private var containerFill: Color {
-        ShellPalette.sidebarMutedInk.opacity(isSelected ? 0.105 : 0.075)
-    }
-
-    private var primaryPaneFill: Color {
-        isSelected ? ShellPalette.accent.opacity(0.82) : ShellPalette.sidebarMutedInk.opacity(0.38)
-    }
-
-    /// Single-pane topology has no focus-within-split to mark, so selection is
-    /// conveyed by the row surface — not focus indigo. A neutral ink, a touch
-    /// stronger when selected, keeps the action/focus accents scarce
-    /// (design-language.md principle 3).
-    private var singlePaneFill: Color {
-        ShellPalette.sidebarInk.opacity(isSelected ? 0.55 : 0.38)
-    }
-
-    private func paneFill(isFocused: Bool) -> Color {
-        if isFocused {
-            return ShellPalette.accent.opacity(isSelected ? 0.88 : 0.78)
-        }
-        return ShellPalette.sidebarMutedInk.opacity(isSelected ? 0.40 : 0.32)
-    }
-
-    private var complexButton: some View {
-        Button {
-            onFocusNextSplitPane(summary)
-        } label: {
-            complexCountOverlay
-        }
-        .buttonStyle(.plain)
-        .help("Focus next split pane")
-        .accessibilityLabel(splitAccessibilityLabel)
-    }
-
-    private var complexCountOverlay: some View {
-        indicatorFrame {
-            ZStack {
-                RoundedRectangle(cornerRadius: ShellRadii.micro, style: .continuous)
-                    .fill(primaryPaneFill)
-                    .frame(maxWidth: .infinity, maxHeight: .infinity)
-
-                Text("\(summary.paneCount)")
-                    .font(.system(size: 8.5, weight: .bold, design: .monospaced))
-                    .foregroundStyle(complexCountForeground)
-            }
-        }
-    }
-
-    private var complexCountForeground: Color {
-        isSelected ? Color.white.opacity(0.92) : ShellPalette.sidebarInk.opacity(0.76)
-    }
-
-    private func segmentButton(paneID: String) -> some View {
-        Button {
-            onFocusSplitPane(paneID)
-        } label: {
-            segmentView(paneID: paneID)
-                .frame(maxWidth: .infinity, maxHeight: .infinity)
-                .contentShape(Rectangle())
-        }
-        .buttonStyle(.plain)
-        .help(summary.focusedPaneID == paneID ? "Focused split pane" : "Focus split pane")
-        .accessibilityLabel(summary.focusedPaneID == paneID ? "Focused split pane" : "Focus split pane")
-    }
-
-    private func segmentView(paneID: String) -> some View {
-        let isFocused = summary.focusedPaneID == paneID
-
-        return RoundedRectangle(cornerRadius: ShellRadii.micro, style: .continuous)
-            .fill(paneFill(isFocused: isFocused))
-    }
-
-    private var splitAccessibilityLabel: String {
-        "Split tab, \(summary.accessibilityTopologyLabel)"
-    }
-}
-
-private struct ShellCompactEmptyAction: View {
-    @Environment(\.accessibilityReduceMotion) private var reduceMotion
-    @FocusState private var isKeyboardFocused: Bool
-    @State private var isHovered = false
-    let title: String
-    let systemImage: String
-    let action: () -> Void
-
-    var body: some View {
-        Button(action: action) {
-            HStack(spacing: 8) {
-                Image(systemName: systemImage)
-                    .font(.system(size: 15, weight: .regular))
-                    .frame(width: ShellSidebarRowMetrics.leadingSlot, height: ShellSidebarRowMetrics.leadingSlot)
-                Text(title)
-                    .font(.system(size: ShellSidebarTypography.titleSize, weight: .regular))
-                Spacer(minLength: 0)
-            }
-            .foregroundStyle(foreground)
-            .padding(.horizontal, ShellSidebarRowMetrics.horizontalInset)
-            .frame(minHeight: ShellSidebarRowMetrics.height)
-            .frame(maxWidth: .infinity, alignment: .leading)
-            .background(
-                ShellSidebarRowBackground(state: visualState)
-            )
-            .contentShape(Rectangle())
-        }
-        .buttonStyle(.plain)
-        .focusable()
-        .focused($isKeyboardFocused)
-        .focusEffectDisabled()
-        .onHover { isHovered = $0 }
-        .animation(reduceMotion ? nil : .easeOut(duration: 0.14), value: visualState)
-        .accessibilityLabel(title)
-    }
-
-    private var visualState: ShellSidebarRowVisualState {
-        isHovered || isKeyboardFocused ? .hover : .normal
-    }
-
-    private var foreground: Color {
-        isHovered || isKeyboardFocused
-            ? ShellPalette.sidebarMutedInk.opacity(0.86)
-            : ShellPalette.sidebarMutedInk.opacity(0.58)
-    }
-}
 #endif
